+++ CHANGELOG.rst
@@ -2,6 +2,26 @@ Changelog
 #########
 
 
+*latest*
+--------
+
+- ``get_receiver`` can now take entire ``Field`` instances, and returns in that
+  case (``fx``, ``fy``, ``fz``) at receiver locations.
+
+- Various small things:
+
+    - Strength input for ``get_source_field`` can now be complex; it also
+      stores now the source location and its strength and moment.
+    - New attribute ``Field.is_electric``, so the field knows if it is electric
+      or magnetic.
+    - Simple ``__repr__`` for ``TensorMesh``, ``Model``, ``Fourier``, ``Time``.
+    - Python 3.8 is now tested.
+
+- Bugfixes:
+
+  - Related to ``get_hx_h0``, ``data_write``, printing in ``Fourier``.
+
+
 *v0.9.1* : VolumeModel
 ----------------------